add an alias for 'add' to delta man page.
 added info (IN COMMENTS) about the bug that park yields an error
when one of the park files has conflicts when unparking.
The reason its in comments is because
i'm not sure if we want to users to see this right now.  i'm not sure
the format for presenting these bugs.  lm or awc should look.

bk: 3a8dbec1GBv2ixzV1cr6WUYtDtqzdA