Add wonderful feature of the week!
Now anytime a rev is passes on the commandline, @csetrev, @tag, or @csetkey
can be used intead.  This allows files to be referenced by tags in
commands line diffs and prs.

This feature uses a cache slices of the ChangeSet file to make it very fast.

bk: 3d41b595cExKiDDU63PMvkuzZQ43cw