Bug 2002-03-01-004
Changed the weighting for the top portion of the bug form so that
it resizes properly when the GUI grows or shrinks. This allows 
all controls and entry fields to be accessible, even with very
odd sized screens (such as the 1024x480 display mentioned in the
bug report)

This is one of those facinating bugs where literally one byte made
a big difference. I love those.  

bk: 3d6f0cce9UKe8ViMXFO7BTTogXebGQ