This fixes one bug and adds one new feature.
Bug: citool now will always report any output by the
commit process, whether it came on stdout or stderr,
and no matter what the exit code. Prior to this fix
it was possible for some messages from the commit 
process to be supressed.

New feature: if 'bk commit' exits with an error code
of 100, citool will exit. If commit exits with any 
other non-zero exit code citool will not exit, forcing
the user to manually exit the program. The reason for
this feature was to allow triggers to do checks for
various things (such as requiring a bug id) without 
requiring the user to restart citool.

bk: 3ddabb1c13cw34n6YDhHCAJJ2fuGWA