Fixes to work around a bug in csettool on Windows. The problem was discovered by Maxtor (a potential customer) who noticed that double clicking on a file would give a tcl error. This only seems to happen on Windows, though I'm not sure why. Perhaps it's related to the way tk recognizes double-click events differently on that platform.
Anyway, this changes how single and double-clicks are processed. The
end result is a) no more tcl error, b) a perceived performance 
improvement (we update the display sooner after clicking to make
the GUI seem more responsive). This latter was because my windows 
box is dreadfully slow and I got tired of waiting 1 or 2 seconds to 
know if my click was received.

bk: 3f721341gTwYoOZ1EcPT4LVw0g5v4w