BUGID: 2002-09-18-001
When automatically including or excluding additional files, citool was
searching the list of files for the filename, pulling out the version
and doing the work based on the version number. However, the search
wasn't specific enough so a search for "foo.c" would also match
"foo.c.save".

This fix adds an additional check that compares each file to the original
included/excluded file to handle this case. 

bk: 3f7af8c0aLT_wU1Owl7qS-cGCSBLlg