Fix some strange error handling with clone that have been reported lately.
First we had a case were the serverInfoBlock was truncated
but clone didn't notice or care.  Now we check for this
case in all network commands.

Next mkdirp() was failing to report the correct errno when
it failed and that could lead to error messages about file
not found instead of permission denyed.
  (ex bk clone URL /nowritepermission/repo)

bk: 3f7c9ad5V-RL1Zy-mT4Y5qlEx6FkXw