Add a new database file BitKeeper/etc/bkstate that maintains per-cset information that is usually a function of the current repository and is unlikely to change frequently. Currently only the bkversion and license is saved.  This is to release early to make the new license server easier to implement.
The bkstate file is special because it always automerges and
is created with a special rootkey so that multiple files created
in parallel will merge to the same inode.

This cset just starts recording the information, later versions
with add code to efficiently access the data.

bk: 3fde2e7bT28007NQGXwkox4oZjIbyw