Snapshot of new tcl based installer.
The thing to include in the zip file and run at install time is
"setup.tcl"; it is built from the utils directory from setupmain.tcl
and a couple other pieces.

It doesn't quite yet fully install anything, and there's a weird-ass
bug on window where the "pick a destination" radiobuttons are all
solid blue which makes it impossible to see which one is selected.

bk: 402c06fd2lFS3E-9ECxrNq_UydfjBg