Fix a number of compiler warnings and missing prototypes on Windows.
Remove cases where file handles are explicitly set to binary mode
on Windows when the handles are already opened in binary by default.
(It is confusing to see code that appears to be needed, but isn't.)

Fix problems with build/buy in tcltk and msys.  We could install the wrong
version when the local repo was missing and we are not working at the 
remote tip.

Fix tcltk/tcl install target where it ends up with two different tclsh.exe
binaries.

bk: 403cec2atiI32knja-ntxMD9Ra6lRQ