A set of changes to how bitkeeper send email for sendbug and support.
We now support sending email to a BK URL that using a new bkd 'MAIL'
command. That is the same as the SMTP 'MAIL' statement and the rest
of the transaction follows the SMTP protocol.  For 3.2.0 the default URL
is a CGI script on bitmover than handles the protocol.

Also remove support.c, it was just a clone of sendbug.c and it made more
sense to just extend sendbug.c to handle support requests as well.

bk: 409919d4d9uVtYAjU8w5XzGc2hq2BQ