There are two cases to worry about when considering .bk_skip files: the "objdir" case, where there is no SCCS subdir, and the "srcdir" case, where there is an SCCS subdir.  Prior to this cset, BK did not differentiate:  any directory with a .bk_skip file was skipped.
This is good for objdirs, but it turns out that a .bk_skip
file in a srcdir causes a lot of higher-level code (check,
pull, push, fix -c) to fail in mysterious ways, and the error
output is confusing and does not point to the root cause.

With this cset, objdirs continue to be handled the same.  But
a srcdir (a directory containing both .bk_skip and SCCS/) is
warned about, and then the .bk_skip is ignored -- the srcdir
is processed as normal, on the assumption that it is a user
error.

The plan is that, if a user actually wants to do this, we can
add code to check for .bk_skip && SCCS/.bk_skip, and handle
it then; Larry said "that is a new interface and then we have
to support it forever".  So there's a comment before
sfiles_skipdir, but I didn't include the code in this cset.

The workaround is to touch .bk_skip and rename SCCS (or move
it into a subdirectory).

BUGID: 2004-01-16-001

bk: 40bf91ffW5PLWdpFsqKgy-b7CPuR2A