Clean up error checking for clone.  If the user specifies an unwritable destination, or no destination (implicit .) in an unwritable directory, detect it right away and error out. Also, correct a few odd bits in how a computed destination is error checked.
In support of the above, implement check_mkdirp, which uses
access(2) to try to answer the question "will mkdirp succeed,
given this path?"  It's, in general, impossible to answer such
a question, but testing with access(2) is fairly robust, and
should never give a false negative.

The traditional UNIX way to do such a thing is "do it, and if it
fails, report the error."  But I shouldn't mkdirp() before I talk
to the bkd (leaving a destination directory around on error exit
would be bad hygiene), and I strongly desire not to talk to the
bkd if the destination is nonwritable.

Add regression tests checking many of these cases in t.clone.

BUGID: 2003-10-02-002 (partial)

bk: 40cf682cwKkYaWD7kRyjGeZTYT1WlQ