Fix the default Cut, Copy, and Paste mechanisms to be more like what the users expect. Also, interact better with the system's clipboard.
Make the 'no files to check-in' errors appear in a message box
instead of in stdout. Windows users that run citool from shellx
don't really have a stdout.

There is still a bug found by Amy: if you select something with the
mouse and then click (thus clearing the selection) and then try to
paste using the middle button, there's nothing to paste!

This is actually a bug with Tk (or a feature?). If you try the same
thing with an xterm, it works as expected. If you run the following
trivial Tk program:

  text .t
  pack .t
  focus .t

and try to do the cut & paste that way, it doesn't work.

According to Bryan that's actually a feature of Tk. Well, I guess
one person's bug is another person's feature ;-)

bk: 41a27c85BdUOj6-nL54JN6wDCDLvAA