* fixes case where citool could save a transient message ("Click Discard   again if you really want to ...") in the c.file. * added test cases for above * added code to guitest.tcl that makes it possible to simulate control keys
This regression snuck in when the code to save comments was revamped. I
stumbled on this bug in response to a customer (savaje) problem where
pressing the discard button then doing control-l put the GUI in a
weird state. The comment-saving revamp made the control-l problem go
away but introduced this new problem.

You can see the bug by backing out the change to citool.tcl, clicking
on a file, click on the Discard button once, then pressing control-n
or control-p to move to a different file. The act of switching files
caused the transient "Click Discard again..." to be saved as the
comments for that file. 

bk: 421e6780yUcK1T780UofhWjlDQqRXQ