Handle the abnormal termination of sfiles in a 'bk -r' command more gracefully. Now you can actually Ctrl-C a bk -r check -avg and not get any false errors.
Also make waitpid support WNOHANG on Windows. 

NOTE FOR 3.3.x MERGE: This cset introduces an interface change to 
sfileDone(). It fixes all instances in this codeline, but if there
are more calls to sfileDone() in 3.3.x that were not in 3.2.x they
also need to be fixed. Having said that, not checking for the return
value of sfileDone is not fatal, it just means the caller doesn't 
notice that 'bk sfiles' exited abnormally.

bk: 42766d50MSzi6hCzJwgLFiQhSAO73g