Better handling of filenames with spaces in resolve, and GUIs.
Fixed some Windows problems with resolve options.

The 'sd' option still dies miserably on Windows, but that is 
unrelated to the files with spaces, so I'm leaving it for another
ChangeSet.

This cset removes _almost_ all calls to spawnvp. When pulling to 3.3.x
the rest of the calls should be removed and the spawnvp() interface
should also be removed.

bk: 42d40eedEYaGqyP_KAX5-PhdwykEzA