Change handling of expired licenses in bk.
With previous versions, an expired eval license is all that
is needed to get unlimited read access to a repository.
This cset attempt to close that hole.

Now leases can no longer be created with expired licenses
and so all read access outside of BitKeeper/etc is disabled.

bk: 42eb6177DSPWJUDRPWLAAryK6maFQA