Change 'bk upgrade' to require a real license key be available before it starts a download.  And then use inskeys to insert that include into the install image that was just downloaded.
This csets makes the inskeys command internal to bk so
that it can be called directly from upgrade,  it
is still build as a seperate standalone command in
addition.

This cset changes inskeys to use more of the functions
provided by our local libc and added a memmem() function
to our local string library since that was useful.

bk: 4474e1c1jJabme9F2gHaiQToKFU2eA