Fix a couple of installer bugs.
If the `bk bin`/config file was read-only, we copied it read only
and then, uh, couldn't write to it.

When things went wrong proceeses would die and Tcl would inform the
user. Use catches to make it shut up.

If you didn't have a network connection and you reconnected to the
network midway in the install, the installer would go crazy.

bk: 4489daaaINAWDaTgCM8FWASsdyZCHQ