This cset adds support for EOLN_WINDOWS which is \r\n on all platforms.  We present a model to the user which is three flags: EOLN_{UNIX,WINDOWS,NATIVE}.  Any of the flags may be set or unset and the code does the right thing, i.e., you set one and the other two are turned off.
The code is weird because EOLN_NATIVE was used to mean native, or if this
bit is off, unix.  These bits are stored on disk and passed on the wire
so adding two new bits, one for windows and one for unix, would break
everyone and force a flag day.  So instead, we add just one new bit,
EOLN_WINDOWS, and do all the dancing in the code to make it appear like
three on the command line.  

xflags.c got a bit stricter about what it likes/doesn't like, it insists
that the xflags recorded match the comments recorded so each time a flag
is changed there needs to be a "Turn on XXX flag" or "Turn off XXX flag"
comment.  The exception is EOLN_UNIX, that is always expressed terms of
on/off for EOLN_NATIVE so xflags.c doesn't have the same hacky code that
the other places do to emulated 3 bits from 2.

bk: 44a48921kSL5q1pUd1uWRWiptLWG_w