Fix getrealname to not call GetLongPathName which returns, sometimes, full paths when you hand it a relative path.  That makes BK very unhappy.
Performance seems slightly better with this code when I did a bk -r edit
in the mysql tree and then timed 5 runs of bk sfiles -E I got an average
of 9.1 seconds; the original code was 9.8.  Regressions were 1.1% slower
and think that's in the noise.

bk: 44db6234SpnOz6cut9oI0YR6ym3pbw