Change the way pclose() and system() return the exit status from a pipeline of commands.  Previously both commands returned the exit status of the last command in the pipeline and failures eariler in the pipeline were ignored.
Now the last non-zero exit status is returned.  So that if the last
command succeeds, but a command eariler in the pipeline failed the
overall command will still return a non-zero exit status.

This makes it harder to overlook failures.
This change is needed for binpool.

bk: 457f066a6Kvp_RFRd2ockLUS2sO0mA