Fix a problem with the previous SFIO mode.
Before when creating a SFIO archive for a binpool data file it would
calculate the actual checksum and compare it to the checksum in the
filename.  If they didn't match then SFIO would exit early with a
truncated file record.  This caused the receving side to generate more
spurious errors when it unpacked the truncated file.

Instead we bypass the actual checksum calculation on the archive creation
side and just attach the alder32 from the filename.  Then the receiving
side will detect the corruption when validating checksums and generate
a cleaner error message.

This optimizes for the common case of having no errors.

bk: 45a64c5c8lEFqvODRsLBb2t93B_YCg