Make it easier to swap Tcl/Tk versions.
All it takes now is a change to the URLs and Keys
and putting the right version in src/gui/tcltk/Makefile

Note that even though the version of TCL_LIBRARYPATH and
TK_LIBRARYPATH are wrong in bk.c, both Tcl and Tk have code
to work around wrong versions and fix them... so even if we
forget to change it in bk.c it should still work.

Also don't set (TCL|TK)_LIBRARY on the Aqua code path. It's not needed and it
interferes with normal Aqua bundle logic.

bk: 45d24988R2k8Nsf-1dHa6z7_Eqzv2w