Extend Rob's dspec parser to ignore whitespace inside condition expressions.  So now:    $if(:GFILE:=ChangeSet) and    $if(:GFILE: = ChangeSet) are equivalent.  Originally the second one included the spaces in the strings on either side of the expression.
Related is that the && and || operators no longer require to be surounded
by spaces to parse correctly.
---
Regressions:
  Fixup a Macos regression failure.
  Fix Rob's testcases to work correctly on my machine.
  Make testcase pass on windows by avoiding quoting issues in the shell
  use old(er)-perl consistantly
---
Since this cset is added all new dspec parsing code Rick
suggested I move it to a new file.  This contains the code to
do that move.  Ideally kw2val() should also move to dspec.c, but
it has to much history to waste and uses lots of slib.c private
functions.
---
Add documentation for =~ and remove
support for !~

bk: 45f568beDnhJ-i8AITXJvhdbIH2_qA