Add a testcase to verify that the backup lease server is up and responding.
Add some error checking for the code in uniq_open() that
rebuilds the keycache when it appears corrupted.  We
have a customer that has a loop where the cache is regenerated
over and over again.

With this fix an error message will be printed, but then the
operation will continue.  To fail delta and commit would be
a more involved change and this shouldn't really happen. ;-)

bk: 46ba6adfvyPAi5bEnIpeoJbFG5j5Lw