Fix and regression for 2007-06-20-001.
bk sfiles -v would die if it was looking at a binary file under revision
control AND fopen on that file failed.

The bug filer had a situation where that binary file was held open un-shared
on Windows causing the fopen() to fail.  The regression induces the failure
more generally by chown'ing the g-file to zero.

Extend the lock command with a -k meaning keepOpen,
hmm, maybe better as -O or ?  Anyway, wanted to try
this out.  Useful for testing on windows.

bk: 46c35e31VxRCZTc3Ib9WH3LJSt_59A