Add a testcase that demonstrates the bug seen at <redacted> this week.  A bk-3.2.x client makes a request to a bk-4.0.x bkd.  The request is properly signed but 3.2.x put the license key in the web headers.  The 4.0.x bkd signs that license key in a lease and trys to use it.
A non-airgap license is really just the customer identifer
that is used to obtain leases from the leaseDB.  The lease
contains a different licence key that encodes the customer's
EULA, options, term, etc...  The invoice number in the two keys
relate each other.  When the identifer gets put into a lease
directly then all the options for that user gets lost.

At <redacted> this causes a user to appear to be using the BKCL EULA
so the bkd would start prompting for that EULA, since the user
normally get the Pro EULA they never agreed to it normally.

In bk-4.0.x the BK_LICENSE header in bkd requests is actually the
license key from the current lease that contains the correct
license terms for this user and that one can be used to create
a new lease.

We avoid this problem by use ignoring BK_LICENSE created by a
bk-3.2.x client.  Technically we could use a license with the
airgap bit set, but that is not worth bothering with.

bk: 46e03843qAEng1n270ei-0XGZ4SD4A