Pass the rootkey in the environment as part of the BK protocol.
This is a change needed in support of remembering repo identities.
The idea is that at some point we'll have a mechanism for pulling
and pushing from repos that have been newrooted via remembering all
the different rootkeys the repos have had. 

We don't have that yet, but this makes sure that old binaries give us
something to work with once we do have it.

Re-enable a bunch of triggers testing that accidentally
was disabled during global trigger developement.

bk: 46f24b55n4UM7PNl-dNWpqpYWe61oQ