Martin Dorey at Bluearc's bug sent in Friday Nov 2, 2007, but with no bug number because the bugdb is down.
A long comment line of exactly the split length would get
split into two lines: the line and a blank line.  This
means that a file processed by the splitting code would
gain a new line each time it was processed.  The commit
code processes a comment twice: initially and after a
pre-commit trigger.  A new line was getting added.
This is more surgical, less elegant fixing of the problem.

bk: 47347f02TYyBnGwgi_ecDYocclkICQ