Fix a mismatch parens problem in slib.c:sccs_delta()
Add a test to verify that BK will complain if a user attempts to delta a
file that has been checked out for editing but changed to read-only.

The test was added because the parens fix means that a block of code
that previously did not get executed now can.

bk: 473496b7YMNW7MWXqxb6ailmwI5PVg