BK currently contains code to make ssh connections with -C for compression if it can.  But then it should disable additional compression.
This code was broken. Sometimes we would compress again to
send over the compressed connection, and other times we
would disable compression when we didn't use a compressed
stream.

This code adds a compressed flag to the remote struct so
that we can know of the bkd channel is already being compressed.
Then compression can be disabled for already compressed
connections.  Also compression of BAM connections are now
controlled.

bk: 478cde8dn4E7qOkf87BcziYDZK4slA