The repo without this cset failed on work because it adds a running of perl to the script, which puts it in the bash hash cache, which means that 'type perl' returns ... hash (/usr/bin/perl) which messes up the test. Replace with the not-known-if-portable type -p which works on work.  bk which doesn't work because of the faked paths, it winds up with one with a space in the name.
Make bam triggers its own test, since paths in a push and pull
are different depending on the presense of bam data or not.
XXX: I haven't yet walked the triggers to have them all be
meaningful tests in a bam context.  That's for after lunch :)

bk: 47979272v4EzJPRftk3m2hkfVR8Mwg