Fix a subtle win32-only bug with the random number generator.
We would use /dev/random (or CSP on win32) to generate the seed
for our pseudo random number generator, but that was too slow
for some operations.  So when bk calls itself we pass the seed
down in the environment.

To prevent people from overriding or interferring with this
seed we would encrypt it and include the current time and
process id.  (This is also used to validate 'restricted' commands.)

On Window we don't have a getppid() call so we couldn't validate that
the RANDSEED in our environment actually came from our parent.  So
if bk calls sh, which then calls multiple bk commands each bk would
see and use the same seed.

The fix is to only pass RANDSEED to a child process if the process is
a call to 'bk'.  Not having RANDSEED set will just cause bk to
regenerate it.

bk: 47d1a0c0Dha461IG8V7H6r-kmI-6Mw