Fix a bug reported by Brennan
If undo restores a file from the deleted directory when running
in checkout:get and partial_check mode then the file will not
be checked out after the undo.  This is because proj_saveCO() will
record the checkout state of the file in the deleted dir (none) and
then proj_restoreAll() will do its thing.

Normally for users this won't be noticed because the consistancy
check will checkout the file, but we were not enabling that chk_gfile()
function in partial_check mode.

I could not see any reason for that so I changed check.c to always
run the chk_gfile() code to make sure files match the checkout mode.

It will still be the case that in checkout:last a file moved from
deleted will not be checked out.   This may surprise users.

bk: 47fb721at47NWpy32WuFZFkbaYNXuQ