NESTED: implement rclone (based on ob's push).
We know we are in an ensemble so we drop right into an
iterator and start cloning.  The other side needs to
start a transaction and send it back; we fake it for now.
The other side will eventually stall the check until the
end but for now we punt on that as well.

bk: 48173b407mmCagKzbDrBWGP6C2qcNg