NESTED: sparse push.
Quite a bit of wackage to handle a sparse push where the other side
has a superset of the repos that participate in the push.  In other
words, the sending side is missing a component that had stuff to be
sent and the receiving side has that component.  In that case, we 
have to fail the push.

They can fix the problem by running populate from someplace with
the missing component[s].

bk: 4817aef8vPyNXB2Rq8Xjs8ymPyxasA