BUGFIX: 'bk clone <repo> .' when run inside a project was doing the clone in the wrong place.
make sane() save/restore the current working directory. It had a side
effect of doing a proj_cd2root(). This was causing "bk clone <repo> ." when
run in an empty directory inside a repo to go back to the repo's root and
do the clone there (on top of the existing repo).

bk: 482b6070yoSV_IBTmRlk06FgSOqXiA