Fix a BAM bug in bp_updateServer() where we would incorrectly make a write locked connection on every null update.
Change BAM server locking to allow commands to wait 30 seconds
for a lock to expire before we fail the operation.  This should
help with contention.

Also include the error messages in the case were locking causes
a failure.

These tests didn't work on some platforms because we didn't always
correctly propagate the exit status from some bk-remote commands
back to the caller.  That is now fixed.

bk: 484700dcgAaNCuljLaTlvidr-MZ7yg