NESTED: make populate a standalone command.
The main difference is that now ensemble_list is run locally
and so the missing components are fetched at the tips that
are expected in the local product cset file.

Also now support a list of URLS:
  bk populate [-M<modules>] url1 url2 ...

fetch from each url until the missing components are found.

bk: 4862f04ajrdkPGwsXQ52rI_WG_4iiw