GONE: make it possible to have a different gone file.
This cset makes it much more possible to use the gone
mechanism to outsource.  You can clone a repo, remove a
pile of stuff, set BK_GONE=BitKeeper/etc/gone-outsource,
run bk -r check -ag | bk gone -, and it works.  All the
contractors need to set BK_GONE but the net result is that
you can feed them a partial tree and pull changes back
from that tree.

BTW, this cset is a great example of why I'm not a big
fan of the ability to concat constant strings.  See the
diffs.

bk: 4891c89ccZedxxhFMjCjwu49ARXlNQ