Make bk notify the explorer.exe plugin about changes.
Whenever the ShellX DLL is mapped into a process (explorer.exe),
it registers itself in the windows registry with a mailslot name.
When BK changes an sfile (e.g. "bk delta") and that change is not
reflected in a change to the gfile (e.g. "bk delta" in checkout:edit),
bk will tell all the registered ShellX DLLs about it via a mailslot.

This is because explorer.exe will miss the change in status on that
file (e.g. from 'modified with diffs' to 'checked in').

Note the code is windows only right now, but it could potentially be
used for "bk explorer" in other platforms.

ENHANCEMENT: Right now, the notifier notifies on a per-file basis. 
It should aggregate all files in each directory and only notify
on a per-directory basis.

RTI: http://work.bitmover.com:2359/cgi-bin/view.cgi?id=2008-08-21-005

bk: 48c8501blAYYrqUbJAAgFOPHuyHDjw