Take advantage of the new fslayer code that can generate a strace of system calls to create a new testcase. t.strace-cnt
This testcase is like t.cset-cnt but it counts the number of
system calls made by bk and generates an error if bk changes.
This is to catch changes that unintentionally hurt performance.

Csets that reduce the number of syscalls will also be flagged.
This is so we will update the baselines won't lose those gains.

The thresholds in the perl script may need to be tweaked if this
gives errors too often.

To regenerate the baseline data after a difference has been determined
to be acceptable, run this:
  STRACE_CNT_SAVE=1 ./doit t.strace-cnt

bk: 491200fd7iIuf-9EuhmLf0yjsP-lSQ