Fix some BAM related bugs when running in an ensemble where we have a mixture of BAM and non-BAM components.
Include a testcase for a bug Rick saw in the dogfood tree.
(The last push one)
The problem was that we were pushing updates to a non-BAM component
when the product contained BAM data while using a file URL.
After the product keysync the client bk had BKD_BAM set in the
environment.  After the second 'bk bkd' is spawned to handle
the non-BAM component push it inherets BKD_BAM and when
takepatch is creating the RESYNC directory sccs_mkroot()
incorrectly created a BAM_MARKER file in the destination
component.  At this point the bkd is inconsistant.
It didn't have a BAM_MARKER at the start of the transaction so
it didn't send BAM=YES to the client, but after takepatch it
did have a BAM_MARKER so it started to send BAM keys.
This caused the client bk to error with:
push failed: recieved BAM keys when not expected

Fixed by moving that BAM_MARKER create from mkroot to clone where
it belonged.

Also fixed BAM server URLs to propagate before BAM data is created
so the new testcases worked and some problems where those
environmental variables can bleed from one component to the next.

And fixed 'bk bam clean' to clean BAM data from a component with
no BAM csets.


bk: 4918942agwRMSd-qEr338GxDZd-1QQ