Add code to correctly handle the component ChangeSet file when running collapse in an ensemble.
When running from the product we strip cset marks, but not
the csets and so leave the components csets pending.

When running from a component, we don't allow the collapse
if the csets to be removed are not already pending.

Add a simple 'bk collapse -d' command to preseve existing 
file deltas.

bk: 49189525mDcjk8F4VN5KQ1By9t5miw