Add code to rset to recurse into nested components if -P is passed. Add code to export to use it by default.
NOTE: bk setup in a product has a bug, so work around it for now.
The bug is that setup in a component without a config file,
because the license (as in Pro, Academic, ...) hard codes
in its deep internals proj = 0, and the marker to mark
this a component hasn't been 'bk new' yet because that needs
a license to do it.  I'd hack something so setup works, except
that there's a larger command line problem.  If we do a 'setup -C',
then we'll get no attach and no config file.  Maybe attach should
zero the config file instead?  Need to think more about this.

bk: 4920f00cnkwrySmzcg7mpG_NeE4qPA