<redacted> hit a case recently where a push would fail with a resolve conflict between two repositories that both pass the consistancy check.
What happened was that a sfile was gone'd and then replaced in the
source repo, but the sfile still existed in the destination.
This cset extends the resolve code so that the messages from bk
would give the user an idea about what is happening and
how to fix it.

bk: 496d0cd9JRifOytY8SjQdMRFcw7BIA