Add a bk extension for Starent.  When running a pull or push we send whole sfiles for any new sfiles or any sfiles with more than 10 deltas. Other than new files this only happens if the local repository has no local csets.
We take advantage of the SFIO added to end of a bk patch for BAM files
and just add the sfiles to that SFIO.  Takepatch doesn't need to know
about this hack at all because sfio will just unpack the files
in place as if takepatch created them.

However the takepatch and makepatch status messages needed to change
as we no longer know how many revisions were transferred.

bk: 49808abcZ0JTovT1NGvWD3Z5jCvPqg