I noticed some cases where proj.c:find_root() can be called over and over because the caching in proj_chdir() didn't really work very well.  Now we just force the last 4 projects to stay open.
Also because we cache more we hit more stale caches when a new
repo is created, so we need to flush in sccs_mkroot() and make
sure the dir->proj mappings are cleared as well.

Fix a pre-exiting proj.c bug that was exposed by this
cset. (p->product)

bk: 49a5fb2bvqJJD-taWZV56S78IcQvwA