blobfs: Add a realBasename() API that returns the real basename of a path.
This fixes the problem of getRealName() not working in a blobfs world
because FindFirstFile() is not mapped by the fslayer. Now getRealName()
on windows uses realBasename() which is handled by the fslayer code.

Note that getRealName() itself cannot be handled by the fslayer since
it deals with paht names. By dealing only with basenames we sidestep
all the issues found when trying to map getRealName() (Thanks Wayne!).

bk: 4a1c8ce4nEl1YSokWwMyoZkOXD4nKw