blobfs: Handle renaming of IDCACHE
In the blobfs code we moved IDCACHE_LOCK to BitKeeper/log because the
windows version of sccs_lockfile() can't handle remapped files.
But this causes a slight compatibility problem so only move the
IDCACHE in remapped repositories and then move the LOCK and the file
together.

bk: 4a3a64bdobn7ebhtN57PQaak8BT0QA