The recent converge.c changes caused a problem with citool when invoked to commit a manual merge after a pull.
The old converge code used to always converge all BitKeeper/etc files
even when no changes occurred so we always got a s.gone file
in RESYNC.  Now the converge code doesn't when there is nothing
to converge, so RESYNC has a gone gfile, but not sfile.
The gfile was added by takepatch.  This extra gfile caused citool
to display it and insist the user deal with it.

Solve this problem by removing the code in takepatch to copy
gone to RESYNC.  And then change loadDB() to automatically handle
reading from the parent repo if gone is missing in RESYNC.

bk: 4a5f266eoAffQ5hy80LHhUAx9whLdw