Fix some races in t.bam-2domains on slow machines. The problem was that in the classic pattern:
	cd S
	bk lock -wt > LOCK &
	bk lock -L
	cd ..
	
	... test ...
	
	bk _kill `cat LOCK`
	rm -rf S

The "bk lock -wt" spawned in the background could still be 
running after the call to "bk _kill" returns and we try to
"rm -rf" the directory. On Windows, you can't remove a directory
if a process is still sitting there, so the regression failed.

This is fixed by introducing a Lock API in setup that besides 
killing the "bk lock" process, it also waits for it.

bk: 4a6e182bSf7spbr-mInm5Ohqjf7DSg