Fix a bug where citool would hang in an infinite update loop if the user tried to scroll while the progress was still being updated.
Fix a bug in commit that would occur if any file was discarded.

Fix a bug where an ignored file was being stored by its full path.

Fix a bug where the ignore file was not being included in the cset
when a file was ignored.

bk: 4a819dd1YrdfWZxme-A8ygShVSDyqA