Fix a bug introduced by the recent bk-nested-bam cset. A lclone of an old-style repo would do a partial upgrade and leave the BAM directory scrambled.
Change the API for bp_serverID() and bp_serverURL() as the
old versions that replied on static varibles were unreliable.
This fixed a bug in clone.c:clone2() where the url from
bp_serverURL() was overwritten when calling bp_serverID().

Also fix that clone2() code to not be called when populating
components.

bk: 4a8c0949NO_HoUwzGH6VE5EOXsfTFA