WINDOWS: access/writable model.
For better or worse, here's the mental model on how we deal with
permissions.  Summary: respect unix style permissions.

By way of motivation, our regression tests do lots of things
like chmod -w and then either expect bk to fail on write perms or
fix them.

We've been using a mix of writable() and access(path, W_OK).
There are some problems with that: the access() approach is somewhat
broken in that if you run as root/Administrator, you will likely be
able to write a file/dir that is marked as read-only.  If you were
to take an old tree and run regressions as root you will see many
of the problems that this series of csets fixes.  In other words,
we were already broken.

writable() is also problematic.  All it does is tell you if there
are any write bits set.  That's nice but if the write bits are
set for the owner only and you are not the owner, it's busted.
We're punting on that problem because windows doesn't really like
the concept of owner/group (their stat doesn't support it).

On windows, until Oscar fixed access(), we were using a model that
simply looked at the permissions and respected them.  Even if you
could write it (due to ACLs), if the perms said no then we said no.

We're going with that model on all platforms.  BitKeeper respects
the permission bits even if you have the ability to override them.
What that means is that places where we used to just call
access(p, W_OK) we now call (writable(p) && (access(p, W_OK) == 0)).

We need access() to be accurate so we can tell if the errors we get
in the face of virus scanners are real or not.  So all/many of the
retry loops will do an access() check after the first failure and
if the access fails we consider it a real error and bail on retrying.

bk: 4a9ec191_WBr2aenlyFnHiew9IwCiw