The commands bk pull -R && bk resolve -a would fail because a pull copies merged (therefore pending) changeset files up to product RESYNC, but the code is in pull_ensemble() and not in resolve, so breaking function into two doesn't do the copy, and ends up with check not passing.
Fix this by moving the existing code from the end of pull.c to
resolve.c.

Also change a product-level resolve to abort if any components
have not yet been resolved.  If we want this can be changed to
actually run the resolves in the components.

Rob:
Fix for when csets_in is updated and for when a component ChangeSet file is
moved up to product's RESYNC dir -- whenever resolve creates a cset, rather
than just a merge cset.

bk: 4a9fbcaeNYQw-LdmX6VACMhl5if6Ug