SFILES: silly me, get rid of a usleep()
Once upon a time, back in the good old days of uniprocessors, this
code seems to be telling us that we starved TK so I decided to 
throw in some sleeps to let TK run.

Taking out the sleep because (a) it slows us way down in large trees
and (b) I tested this by hand on a Thinkpad T23 w/ 512MB with and
without the change and it worked fine.

bk: 4aa8193egJphjwFTqCEeM_IeAjMuxg