R/CLONE: symlink and empty dir handling.
Two client side changes, one to grab the dest from the url (if
possible) in case the URL is a symlink (we typically want the 
named basename not the actual remote dir basename).  The other
is to insist that bk://hostname/path urls always have a path
component so we don't splatter all over an empty dir.  Wayne
didn't like the automagic pick the basename of the source in
this case, said it messes with urllist and ssh proxies.

bk: 4aa82883tOkqQQBdmgNxktqd_aseoQ