Fix 'bk clone' to allow repositories to be cloned underneath existing products.
Many of the problems came down to proj_cd2product() which
would cd out of a standalone repo and into a product that
happens to be above it.  This caused 'bk parent' to set the
parent of the wrong repository.

Also fix clone to no longer just assume -p(no parent) if you
were cloning a repo under a product.  Have attach remove the
parent instead.

bk: 4ab10ceddnoARC2_0lKjfBxDM2N64A