Proposed fix for Coverity's complaint:
User A checks out a file for editing; user B does the delta.  The complaint was
that the delta actually works, the pfile is removed but a writable gfile is left
behind causing later commands to complain about writeable gfile with no pfile.

The fix is to try to re-get the gfile if fix_gmode() fails (and to
remove the perror()).

Testing (in src/t/t.delta) depends on being able to chown to someone else
and is limited to HPUX.

bk: 4ab91fa7plmGByX5RSoU5lmB10VIzA