First pass at implementing locking for nested repositories.
They are described in src/Notes/LOCKING-NESTED which has some
pointers to the wiki describing things in more detail. 

Push was reworked in this cset to behave like pull does (not
doing the product as part of the nested loop, but instead 
wrapping the takepatch/resolve around the loop). It was also
mostly rewritten for clarity. God have mercy on the poor soul
that tries to pull old changes to push.c. 

Since this cset only implements write locks, all 
commands lock only the side that is writing. I.e. clone and pull
lock the local side and push locks the remote side. Rclone
is not implemented yet because it works in a weird way so it
would need a protocol change.

bk: 4ace1e696TOdGeKnVeR61Tue7GmwjQ